+++ Common/src/main/java/fuzs/puzzleslib/impl/core/proxy/SidedProxy.java
@@ -14,12 +14,10 @@ public interface SidedProxy {
 
     void registerEventHandlers();
 
-    @Deprecated(forRemoval = true)
     default Player getClientPlayer() {
         throw new RuntimeException("Client player accessed for wrong side!");
     }
 
-    @Deprecated(forRemoval = true)
     default Level getClientLevel() {
         throw new RuntimeException("Client level accessed for wrong side!");
     }