+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.5.4-1.21.5] - 2025-04-03
+### Add
+- Add `HSV` utility class
+- Add one more `RegistryManager::registerSpawnEggItem` overload
+### Fixed
+- Fix entity attributes failing to register on NeoForge
+
 ## [v21.5.3-1.21.5] - 2025-04-02
 ### Added
 - Add `ItemTooltipRegistry`